refactor: DTO 모두 내장 클래스에서 분리 및 record로 변경
