+++ backend/src/main/java/middle_point_search/backend/domains/room/controller/RoomController.java
@@ -19,12 +19,12 @@ import lombok.RequiredArgsConstructor;
 import middle_point_search.backend.common.dto.DataResponse;
 import middle_point_search.backend.common.dto.ErrorResponse;
 import middle_point_search.backend.common.util.MemberLoader;
-import middle_point_search.backend.domains.room.dto.RoomDTO;
-import middle_point_search.backend.domains.room.dto.RoomDTO.FindRoomDetailResponse;
-import middle_point_search.backend.domains.room.dto.RoomDTO.CreateRoomResponse;
-import middle_point_search.backend.domains.room.dto.RoomDTO.ExistRoomResponse;
-import middle_point_search.backend.domains.room.dto.RoomDTO.UpdateRoomNameRequest;
-import middle_point_search.backend.domains.room.dto.RoomDTO.UpdateRoomMemoRequest;
+import middle_point_search.backend.domains.room.dto.request.CreateRoomRequest;
+import middle_point_search.backend.domains.room.dto.request.UpdateRoomMemoRequest;
+import middle_point_search.backend.domains.room.dto.request.UpdateRoomNameRequest;
+import middle_point_search.backend.domains.room.dto.response.CreateRoomResponse;
+import middle_point_search.backend.domains.room.dto.response.ExistRoomResponse;
+import middle_point_search.backend.domains.room.dto.response.FindRoomDetailResponse;
 import middle_point_search.backend.domains.room.service.RoomService;
 
 @Tag(name = "ROOM API", description = "방에 대한 API입니다.")
@@ -70,7 +70,7 @@ public class RoomController {
 			)
 		}
 	)
-	public ResponseEntity<DataResponse<CreateRoomResponse>> createRoom(@RequestBody @Valid RoomDTO.CreateRoomRequest request) {
+	public ResponseEntity<DataResponse<CreateRoomResponse>> createRoom(@RequestBody @Valid CreateRoomRequest request) {
 		CreateRoomResponse response = roomService.createRoom(request);
 
 		return ResponseEntity.ok(DataResponse.from(response));
