+++ backend/src/main/java/middle_point_search/backend/domains/room/dto/request/UpdateRoomMemoRequest.java
@@ -1,0 +1,6 @@
+package middle_point_search.backend.domains.room.dto.request;
+
+public record UpdateRoomMemoRequest(
+	String memo
+) {
+}
