+++ backend/src/main/java/middle_point_search/backend/domains/room/dto/response/CreateRoomResponse.java
@@ -1,0 +1,9 @@
+package middle_point_search.backend.domains.room.dto.response;
+
+public record CreateRoomResponse(
+	String id
+) {
+	public static CreateRoomResponse from(String id) {
+		return new CreateRoomResponse(id);
+	}
+}
