+++ backend/src/main/java/middle_point_search/backend/domains/room/dto/response/ExistRoomResponse.java
@@ -1,0 +1,9 @@
+package middle_point_search.backend.domains.room.dto.response;
+
+public record ExistRoomResponse(
+	boolean existence
+) {
+	public static ExistRoomResponse from(boolean existence) {
+		return new ExistRoomResponse(existence);
+	}
+}
